refactor: 名前のマッピングメソッドをUserRepositoryに移動
